refactor(lexer): extract inline tests from lexer.rs to tests/lexer.rs
Moves the ~490-line test module from crates/php-lexer/src/lexer.rs into
crates/php-lexer/tests/lexer.rs as an integration test file, reducing the
production source from 1,820 to 1,330 lines. All 32 tests continue to pass.
Adds php-ast as a dev-dependency so the integration test can use Span directly.

Closes #280
